add functions CRUD incidents

+++ src/hooks/useCalendarStore.ts
@@ -1,6 +1,7 @@
+import Swal from "sweetalert2";
 import { incidentApi } from "../apis";
+import { Toast } from "../helpers";
 import { eventCalendar, onAddNewEvent, onDeleteEvent, onLoadIncidents, onSetActiveEvent, onUpdateEvent, useAppDispatch, useAppSelector } from "../store";
-
 
 
 
@@ -10,43 +11,69 @@ export const useCalendarStore = () => {
     const dispatch = useAppDispatch();
 
 
-    const setEventActive = (evento: eventCalendar) => dispatch( onSetActiveEvent(evento) )
+    const setEventActive = (evento: eventCalendar) => dispatch(onSetActiveEvent(evento))
 
 
     const startSavingEvent = async (calendarEvent: Partial<eventCalendar>) => {
-        
-        if(calendarEvent._id) {
-            dispatch( onUpdateEvent( calendarEvent ) )  
-        }else{
 
-            try {
-                const { data: incident } = await incidentApi.post('/incident', calendarEvent);
-                dispatch( onAddNewEvent({ ...incident }) )  
-            } catch (error) {
-                console.log(error);
+
+        try {
+
+            if (calendarEvent._id) {
+                const { data } = await incidentApi.patch(`/incident/${calendarEvent._id}`, calendarEvent);
+                dispatch(onUpdateEvent({ ...data }));
+                Toast.fire({ icon: "success", title: "Saved event" });
+                return;
             }
-            
+
+
+            const { data: incident } = await incidentApi.post('/incident', calendarEvent);
+            dispatch(onAddNewEvent({ ...incident }));
+            Toast.fire({ icon: "success", title: "Saved event" });
+
+        } catch (error: any) {
+            console.log(error);
+            Toast.fire({ icon: "error", title: error.response.data.message });
         }
     }
 
     const startDeletingEvent = () => {
-        dispatch( onDeleteEvent() );
+        Swal.fire({
+            title: "Delete event '" + activeEvent?.title + "' ?",
+            text: "Are you sure you want to delete the selected event?",
+            icon: "question"
+        }).then(async (result) => {
+            /* Read more about isConfirmed, isDenied below */
+            if (result.isConfirmed) {
+
+                try {
+                    await incidentApi.delete(`/incident/${activeEvent!._id}`);
+                    dispatch(onDeleteEvent());
+                    Toast.fire({ icon: "success", title: "event deleted" });
+                } catch (error: any) {
+                    Toast.fire({ icon: "error", title: error.response.data.message });
+                }
+
+            }
+        });
+
+
     }
 
 
     const startLoadingIncident = async () => {
         try {
             const { data: incidents } = await incidentApi.get('/incident');
-            dispatch( onLoadIncidents(incidents) )  
+            dispatch(onLoadIncidents(incidents))
         } catch (error) {
             console.log(error);
         }
     }
 
-    
+
     return {
         //Propiedades
-        events, 
+        events,
         activeEvent,
         hasEventSelected: !!activeEvent,
